Simplify running of release scripts with dependency groups
While verifying the release, I realised that there were a small
issue with setting up the environment - the installation instructions
asked to install burr without extras in local dev env, but that
was not enoughm you need to also install `cli` extra.

I have updated the docs appropriately, but also added an alternative
workflow for running the installation scripts with `uv` and the
standard "dependency groups". The dependency groups are
already standardised and used by `uv` and others - for example to
install dev dependency group automatically when `uv sync` is used.

This makes it as easy to run the scripts as possible without having
all the venv manual installation and syncing (uv does it for you).